chore: extract commit from tag retrieval to function for clarity; move commit finder tests to new file
